Add interactive mode with keyboard commands (#326)
* Implement interactive mode with keyboard commands

* Ask people to reopen the app when toggling dev mode

This is currently required for the change to take effect,
see: https://github.com/expo/expo/issues/413

* Condense instructions to take less vertical space

* Hide the iOS emulator option on Windows

* Highlight the 'R' keyboard shortcut

* Add `--no-interactive` CLI argument to disable the interactive mode

`npm start -- --no-interactive` runs the server without interactivity.

* Fix interactive mode input on Windows

The `readline` setup on Windows that nicely suppresses the annoying
"Terminate batch job (Y/N)" prompt when pressing Ctrl+c also caused
the keyboard commands on interactive mode to be printed to stdout.
Fix this by piping the `readline` output to a MuteStream.

* Stop the packager gracefully when exiting interactive mode

* Format new code with Prettier

* Do not create a readline interface in the interactive mode

Creating the readline interface on Windows to listen for Ctrl+c caused
the text "Stopping packager" to be printed twice. In the interactive
mode stopping the packager is handled by the keypress listener set
by the interactive mode, so the readline interface is not necessary.

* Fix bold text style bleeding outside the bold text part on Windows

Work around the bold style not being closed on Windows by adding reset
styles after each bold style.